+++ sphinxcontrib/needs/builder.py
@@ -71,3 +71,17 @@ class NeedsBuilder(Builder):
 
     def get_target_uri(self, docname, typ=None):
         return ""
+
+
+def build_needs_json(app, exception):
+
+    if not app.env.config.needs_build_json:
+        return
+
+    # Do not create an additional needs.json, if builder is already "needs".
+    if isinstance(app.builder, NeedsBuilder):
+        return
+
+    needs_builder = NeedsBuilder(app)
+    needs_builder.set_environment(app.env)
+    needs_builder.finish()