+++ sphinxcontrib/needs/directives/needpie.py
@@ -1,3 +1,4 @@
+import copy
 import os
 
 import matplotlib
@@ -13,6 +14,7 @@ import matplotlib.pyplot
 from docutils.parsers.rst import directives
 
 from sphinxcontrib.needs.logging import get_logger
+from sphinxcontrib.needs.utils import check_and_get_external_filter_func
 
 logger = get_logger(__name__)
 
@@ -43,6 +45,9 @@ class NeedpieDirective(FilterBase):
         "text_color": directives.unchanged_required,
         "shadow": directives.flag,
     }
+
+    # Update the options_spec only with value filter-func defined in the FilterBase class
+    option_spec["filter-func"] = FilterBase.base_option_spec["filter-func"]
 
     def run(self):
         env = self.state.document.settings.env
@@ -94,6 +99,8 @@ class NeedpieDirective(FilterBase):
             "shadow": shadow,
             "text_color": text_color,
         }
+        # update filter-func with needed information defined in FilterBase class
+        env.need_all_needpie[targetid]["filter_func"] = self.collect_filter_attributes()["filter_func"]
 
         return [targetnode] + [Needpie("")]
 
@@ -125,13 +132,53 @@ def process_needpie(app, doctree, fromdocname):
             matplotlib.style.use("default")
 
         content = current_needpie["content"]
+
         sizes = []
-        for line in content:
-            if line.isdigit():
-                sizes.append(float(line))
-            else:
-                result = len(filter_needs(app, app.env.needs_all_needs.values(), line))
-                sizes.append(result)
+        if content and not current_needpie["filter_func"]:
+            for line in content:
+                if line.isdigit():
+                    sizes.append(float(line))
+                else:
+                    result = len(filter_needs(app, app.env.needs_all_needs.values(), line))
+                    sizes.append(result)
+        elif current_needpie["filter_func"] and not content:
+            try:
+                # check and get filter_func
+                filter_func, filter_args = check_and_get_external_filter_func(current_needpie)
+                # execute filter_func code
+                # Provides only a copy of needs to avoid data manipulations.
+                context = {
+                    "needs": copy.deepcopy(list(env.needs_all_needs.values())),
+                    "results": [],
+                }
+                args = []
+                if filter_args:
+                    args = filter_args.split(",")
+                for index, arg in enumerate(args):
+                    # All rgs are strings, but we must transform them to requested type, e.g. 1 -> int, "1" -> str
+                    context[f"arg{index + 1}"] = arg
+
+                if filter_func:
+                    filter_func(**context)
+                sizes = context["results"]
+                # check items in sizes
+                if not isinstance(sizes, list):
+                    logger.error(
+                        f"The returned values from the given filter_func {filter_func.__name__} is not valid."
+                        " It must be a list."
+                    )
+                for item in sizes:
+                    if not isinstance(item, int) and not isinstance(item, float):
+                        logger.error(
+                            f"The returned values from the given filter_func {filter_func.__name__} is not valid. "
+                            "It must be a list with items of type int/float."
+                        )
+            except Exception as e:
+                raise e
+        elif current_needpie["filter_func"] and content:
+            logger.error("filter_func and content can't be used at the same time for needpie.")
+        else:
+            logger.error("Both filter_func and content are not used for needpie.")
 
         labels = current_needpie["labels"]
         if labels is None: