+++ sphinxcontrib/needs/needs.py
@@ -6,7 +6,7 @@ from sphinx.errors import SphinxError
 from sphinx.roles import XRefRole
 
 from sphinxcontrib.needs.api.configuration import add_extra_option
-from sphinxcontrib.needs.builder import NeedsBuilder
+from sphinxcontrib.needs.builder import NeedsBuilder, build_needs_json
 from sphinxcontrib.needs.config import NEEDS_CONFIG
 from sphinxcontrib.needs.defaults import (
     DEFAULT_DIAGRAM_TEMPLATE,
@@ -224,6 +224,8 @@ def setup(app):
 
     app.add_config_value("needs_string_links", {}, "html", types=[dict])
 
+    app.add_config_value("needs_build_json", False, "html", types=[bool])
+
     # Define nodes
     app.add_node(Need, html=(html_visit, html_depart), latex=(latex_visit, latex_depart))
     app.add_node(
@@ -316,6 +318,7 @@ def setup(app):
     app.connect("doctree-resolved", process_need_count)
     app.connect("doctree-resolved", process_need_func)
     app.connect("build-finished", process_warnings)
+    app.connect("build-finished", build_needs_json)
     app.connect("env-updated", install_lib_static_files)
 
     # Called during consistency check, which if after everything got read in.